# Performance improvements to dependency resolution and caching