--- a/gradle-plugin/src/main/kotlin/io/github/slimjar/SlimJarPlugin.kt
+++ b/gradle-plugin/src/main/kotlin/io/github/slimjar/SlimJarPlugin.kt
@@ -26,15 +26,16 @@ package io.github.slimjar
 
 import com.github.jengelman.gradle.plugins.shadow.tasks.ShadowJar
 import io.github.slimjar.exceptions.ShadowNotFoundException
-import io.github.slimjar.func.*
+import io.github.slimjar.func.applyReleaseRepo
+import io.github.slimjar.func.applySnapshotRepo
+import io.github.slimjar.func.createConfig
 import io.github.slimjar.task.SlimJar
 import org.gradle.api.Plugin
 import org.gradle.api.Project
-import org.gradle.api.artifacts.dsl.DependencyHandler
 import org.gradle.api.plugins.JavaPlugin
-import org.gradle.kotlin.dsl.DependencyHandlerScope
 import org.gradle.kotlin.dsl.extra
 import org.gradle.kotlin.dsl.maven
+
 const val SLIM_CONFIGURATION_NAME = "slim"
 const val SLIM_API_CONFIGURATION_NAME = "slimApi"
 const val SLIM_JAR_TASK_NAME = "slimJar"
@@ -51,9 +52,17 @@ class SlimJarPlugin : Plugin<Project> {
             throw ShadowNotFoundException("SlimJar depends on the Shadow plugin, please apply the plugin. For more information visit: https://imperceptiblethoughts.com/shadow/")
         }
 
-        val slimConfig = createConfig(SLIM_CONFIGURATION_NAME, JavaPlugin.COMPILE_ONLY_CONFIGURATION_NAME, JavaPlugin.TEST_IMPLEMENTATION_CONFIGURATION_NAME)
+        val slimConfig = createConfig(
+            SLIM_CONFIGURATION_NAME,
+            JavaPlugin.COMPILE_ONLY_CONFIGURATION_NAME,
+            JavaPlugin.TEST_IMPLEMENTATION_CONFIGURATION_NAME
+        )
         if (plugins.hasPlugin("java-library")) {
-            createConfig(SLIM_API_CONFIGURATION_NAME, JavaPlugin.COMPILE_ONLY_API_CONFIGURATION_NAME, JavaPlugin.TEST_IMPLEMENTATION_CONFIGURATION_NAME)
+            createConfig(
+                SLIM_API_CONFIGURATION_NAME,
+                JavaPlugin.COMPILE_ONLY_API_CONFIGURATION_NAME,
+                JavaPlugin.TEST_IMPLEMENTATION_CONFIGURATION_NAME
+            )
         }
 
         val slimJar = tasks.create(SLIM_JAR_TASK_NAME, SlimJar::class.java, slimConfig)
@@ -66,7 +75,10 @@ class SlimJarPlugin : Plugin<Project> {
                 repositories.maven("https://repo.vshnv.tech/snapshots/")
             }
         }
-        project.dependencies.extra.set("slimjar", asGroovyClosure("+"){ version -> "io.github.slimjar:slimjar:$version"})
+        project.dependencies.extra.set(
+            "slimjar",
+            asGroovyClosure("+") { version -> "io.github.slimjar:slimjar:$version" }
+        )
         // Hooks into shadow to inject relocations
         val shadowTask = tasks.withType(ShadowJar::class.java).firstOrNull() ?: return
         shadowTask.doFirst {
@@ -77,6 +89,10 @@ class SlimJarPlugin : Plugin<Project> {
                 }
             }
         }
+
+        /*slimJar.outputs.upToDateWhen {
+            true
+        }*/
 
         // Runs the task once resources are being processed to save the json file
         tasks.findByName(RESOURCES_TASK)?.finalizedBy(slimJar)