This PR enables SlimJar to cache data to decrease build time.
Also switching to coroutines over parallelStream decreased resolution even more.

Some time comparisons with this PR using 12 dependencies (and transitive ones):
```yaml
[Clean compile]
1.3.0: 1m 2s
PR: 23s

[Add 1 dependency]
1.3.0: - # Same as clean
PR: 3s

[Remove 1 dependency]
1.3.0: - # Same as clean
PR: 3s

[No dependency change]
1.3.0: - # Same as clean
PR: UP-TO-DATE 5ms
```